+++ bocadillo/__init__.py
@@ -1,5 +1,6 @@
 from .api import API
 from .static import static
 from .middleware import RoutingMiddleware
+from .media import Media
 
 __version__ = '0.5.0'