+++ bocadillo/response.py
@@ -1,29 +1,33 @@
-import json
-from typing import AnyStr
+from typing import AnyStr, Any
 
 from starlette.requests import Request
 from starlette.responses import Response as _Response
+
+from bocadillo.media import Media
 
 
 class Response:
     """Response builder."""
 
-    def __init__(self, request: Request):
+    def __init__(self, request: Request, media: Media):
         self.request = request
         self._content: AnyStr = None
         self.status_code: int = None
         self.headers = {}
+        self._media = media
+
+    def _set_media(self, value: Any, media_type: str):
+        content = self._media.serialize(value, media_type=media_type)
+        self.headers['content-type'] = media_type
+        self._content = content
 
     def __setattr__(self, key, value):
         if key == 'text':
-            self.headers['Content-Type'] = 'text/plain'
-            self._content = value
+            self._set_media(value, media_type=Media.PLAIN_TEXT)
         elif key == 'html':
-            self.headers['Content-Type'] = 'text/html'
-            self._content = value
+            self._set_media(value, media_type=Media.HTML)
         elif key == 'media':
-            self.headers['Content-Type'] = 'application/json'
-            self._content = json.dumps(value)
+            self._set_media(value, media_type=self._media.type)
         elif key == 'content':
             self._content = value
         else:
@@ -35,7 +39,7 @@ class Response:
             self.status_code = 200
 
         if self.status_code != 204:
-            self.headers.setdefault('Content-Type', 'text/plain')
+            self.headers.setdefault('content-type', Media.PLAIN_TEXT)
 
         response = _Response(
             content=self._content,