+++ docs/.vuepress/config.js
@@ -73,6 +73,7 @@ module.exports = {
                 collapsable: false,
                 children: [
                     '/guide/how-to/custom-cli-commands',
+                    '/guide/how-to/extra-media-handlers',
                 ],
             },
             {
@@ -88,6 +89,7 @@ module.exports = {
                     '/guide/api-guides/templates',
                     '/guide/api-guides/static-files',
                     '/guide/api-guides/error-handling',
+                    '/guide/api-guides/media',
                     '/guide/api-guides/middleware',
                     '/guide/api-guides/cli',
                     '/guide/api-guides/cors',